refactor test file names inputs_handler/outputs_handler
