+++ tests/test_input_osm_reader.py
