+++ tests/test_output_sanitiser.py
